Fix /d/relational-database-many-to-many
Basically, the name of the link says "relational databases many-to-many"
(also known as implicit many-to-many), but if you follow it, you end up
on the mongodb many-to-many section. The section we want to link to in
that error message is
https://www.prisma.io/docs/concepts/components/prisma-schema/relations/many-to-many-relations#implicit-many-to-many-relations.

As far as we know, this is only used in engines error messages.

More context on slack: https://prisma-company.slack.com/archives/C5Z9TH6N9/p1649259671210389